% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iopac_wrap.r
\name{iopac_wrap}
\alias{iopac_wrap}
\title{Create all IOPAC multipliers}
\usage{
iopac_wrap(
  impbridgein = impbridgelist,
  costfin = costflist_2020,
  multsin = mults,
  ticsin = ticslist_2020,
  ecpiin = ecpi,
  taxesin = taxes,
  prodflowin = prodflow,
  markupsin = markups_2020,
  fskeyin = fskey,
  comnamesin = comnames,
  procrecin = procrec
)
}
\arguments{
\item{impbridgein}{Commodity flows data from IMPLAN. This is a list
containing data for the "vessel" and "processor". Each list element
is then a data frame with columns for the Commodity code, Type of
commodity, and Share, denoting the share each commodity holds for
their given type. Note that some shares may not sum to unity because
they may be further weighted by additional data.}

\item{costfin}{Cost data from NWFSC surveys. This is a list containing
data for the "vessel" and "processor" in each element. Each list
additional columns for the cost share for each vessel type.}

\item{multsin}{Output, income, and employment multipliers from IMPLAN.
This is a list containing three data frames, with a column for the
Commodity Code, and additional columns for each study area region.}

\item{ticsin}{Fish tickets by revenue from PACFIN. This is a list with a
number of elements equal to the number of study area regions in
mults. Each element is then a data frame with the first column
denoting commodity types, and additional columns representing
vessel types, which must match those in costf.}

\item{ecpiin}{Multipliers for employee compensation and proprietary
income from IMPLAN. This is a data frame with a column Type that
denotes the type of multiplier (of 6 variants), and additional
columns corresponding to study area regions from mults and ticsin.}

\item{taxesin}{Multipliers for induced effects from taxes from IMPLAN.
This is a data frame with rows for output, income, and employment,
and columns for each region. Only data for states (WA, OR, and CA)
and the West Coast are included.}

\item{prodflowin}{A three column data frame denoting product flow. The
first column Region denotes the study area regions matching mults
and ticsin. The other two columns denote total supply and processor
demand for each region.}

\item{fskeyin}{A data frame denoting the relationship between species
(Markups), commodity types (Sector), and various gear types.}

\item{comnamesin}{A data frame denoting the relationship between
commodity types and corresponding commodity type codes, if used.}

\item{procrecin}{A data frame with weights for commodity flows data from
IMPLAN (impbridgein), if used.}

\item{marksupsin}{Processor markups from NWFSC data. This is a data frame
with the first column Markups corresponding to commodity types from
ticsin. Then the column Value denotes processor markups.}
}
\value{
A (n*m) by 12 data frame, where the first three columns
correspond to the geographic region, sector, and name of sector for
a multiplier, where there are \emph{n} regions and \emph{m} sectors.
}
\description{
A high-level wrapper function to create impact multipliers
}
