% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_p_mults.r
\name{make_p_mults}
\alias{make_p_mults}
\title{Create IOPAC processor multipliers}
\usage{
make_p_mults(
  impbridge,
  costf,
  mults,
  type,
  sector,
  ticsin,
  ecpi,
  taxes,
  procrec,
  prodflow,
  markups,
  fskey
)
}
\arguments{
\item{impbridge}{Commodity flows data from IMPLAN. This is a data frame
containing data for the "processor". The data frame has columns for
the Commodity code, Type of commodity, and Share, denoting the share
each commodity holds for their given type. Note that some shares
may not sum to unity because they may be further weighted by
additional data.}

\item{costf}{Cost data from NWFSC surveys. This is a data frame for the
"processor". The data frame contains columns for the Type of cost,
and additional columns for the cost share for each vessel type.}

\item{mults}{Output, income, and employment multipliers from IMPLAN. This
is a data frames corresponding to the chosen type, with a column for
for the Commodity Code, and additional columns for each study area
region.}

\item{type}{String. Type of output ("output", "employment", or "income").}

\item{sector}{String. Commodity type, corresponding to those in ticsin.}

\item{ticsin}{Fish tickets by revenue from PACFIN. This is a data frame
corresponding to the study area chosen in sector. Each data frame
has the first column denoting commodity types, and additional
columns representing vessel types, which must match those in costf.}

\item{ecpi}{Multipliers for employee compensation and proprietary income
from IMPLAN. This is a data frame with a column Type that denotes
the type of multiplier (of 6 variants), and additional columns
corresponding to study area regions from mults and ticsin.}

\item{taxes}{Multipliers for induced effects from taxes from IMPLAN. This
is a data frame with rows for output, income, and employment, and
columns for each region. Only data for states (WA, OR, and CA) and
the west coast are included.}

\item{procrec}{A data frame with weights for commodity flows data from
IMPLAN (impbridge), if used.}

\item{prodflow}{A three column data frame denoting product flow. The
first column Region denotes the study area regions matching mults
and ticsin. The other two columns denote total supply and processor
demand for each region.}

\item{fskey}{A data frame denoting the relationship between species
(Markups), commodity types (Sector), and various gear types.}

\item{marksups}{Processor markups from NWFSC data. This is a data frame
with the first column Markups corresponding to commodity types from
ticsin. Then the column Value denotes processor markups.}
}
\value{
procoutcoef is a 1x1 multiplier.
}
\description{
A high-level wrapper function create impact multipliers
}
