% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_v_mults.r
\name{make_v_mults}
\alias{make_v_mults}
\title{Create IOPAC vessel multipliers}
\usage{
make_v_mults(
  impbridge,
  costf,
  mults,
  type,
  sector,
  ticsin,
  ecpi,
  taxes,
  output = NULL
)
}
\arguments{
\item{impbridge}{Commodity flows data from IMPLAN. This is a data frame
containing data for the "vessel". The data frame has columns for the
Commodity code, Type of commodity, and Share, denoting the share
each commodity holds for their given type. Note that some shares
may not sum to unity because they may be further weighted by
additional data.}

\item{costf}{Cost data from NWFSC surveys. This is a data frame for the
"vessel". The data frame contains columns for the Type of cost, and
additional columns for the cost share for each vessel type.}

\item{mults}{Output, income, and employment multipliers from IMPLAN. This
is a data frames corresponding to the chosen type, with a column for
the Commodity Code, and additional columns for each study area
region.}

\item{type}{String. Type of output ("output", "employment", or "income").}

\item{sector}{String. Commodity type, corresponding to those in ticsin.}

\item{ticsin}{Fish tickets by revenue from PACFIN. This is a data frame
corresponding to the study area chosen in sector. Each data frame
has the first column denoting commodity types, and additional
columns representing vessel types, which must match those in costf.}

\item{ecpi}{Multipliers for employee compensation and proprietary income
from IMPLAN. This is a data frame with a column Type that denotes
the type of multiplier (of 6 variants), and additional columns
corresponding to study area regions from mults and ticsin.}

\item{taxes}{Multipliers for induced effects from taxes from IMPLAN. This
is a data frame with rows for output, income, and employment, and
columns for each region. Only data for states (WA, OR, and CA) and
the west coast are included.}

\item{output}{Type of output from the function. Default output are vessel
level multipliers. "mults" gives industry category specific multipliers.
"diremp" outputs direct employment. "dirinc" outputs direct income.
"propinc" outputs proprietary income. "indiremp" outputs indirect
employment. "indirinc" outputs indirect income.}
}
\value{
vesoutcoef is a 1x1 multiplier.
}
\description{
A function that creates impact multipliers for vessels
}
