% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{prodflow}
\alias{prodflow}
\title{Processor product flow}
\format{
A data.frame with columns:
\describe{
\item{Region}{Study area regions corresponding to fish tickets}
\item{Total supply}{Total supply of product in each region}
\item{Processor demand}{Processor demand for product in each region}
}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{head(prodflow)
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{##       Region Total.Supply Processor.demand
## 1         WC   1023415000       1300548000
## 2 California    394765200        202264900
## 3     Oregon    257551400        180982300
## 4 Washington    371098600        917300800
## 5    Astoria     63881120         59733680
## 6  BodegaBay     16779160          6358780
}\if{html}{\out{</div>}}
}
\usage{
prodflow
}
\description{
A dataset denoting product flow for processors by region. Note these are implan data; Total Supply is equal to Total Commodity Supply for Fish under Trade Flows and Industry/Institution RPC; Processor Demand is equal to Gross Commodity Demand for Sector 93.
}
\keyword{datasets}